requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.2
gunicorn==21.2.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.7
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
whitenoise==6.5.0
